feat: Add panel to editor when file is opened
- Create EditorGroupsPanelBuilder class
- Add method to add panel to editor when file is opened
- Ensure panel is added to editor when file is opened
